pyproject.toml
[project]
name = "mem0-mcp-server"
version = "0.1.0"
description = "Mem0 MCP Server - local deployment with device token auth"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "mem0ai>=1.0.9",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
mem0-mcp-server = "mem0_mcp_server.server:main"
mem-cli = "mem0_mcp_server.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mem0_mcp_server"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true
